requirements.txt
pytelegrambotapi==3.6.3
pysocks==1.6.8
jenkinsapi==0.3.6
python-gitlab==1.4.0
psycopg2-binary